LayoutTests/fast/text-extraction/debug-text-extraction-ignore-autofilled-fields-expected.txt

root
	input,'email',placeholder='Enter your email'
	input,'password',placeholder='Enter your password',secure
	input,'number',placeholder='Security code','123456'
